scripts/requirements.txt
cysystemd
pydantic
python-dotenv
py-cord @ git+https://github.com/Pycord-Development/pycord.git
